[rail_inspector] Extract NewFrameworkDefaultsFile
To make it easier to change what gets passed to the FrameworkDefaults
check.
